Fix build warnings in CTester, AsirikuyCommon, and SymbolAnalyzer
CTesterFrameworkAPI/tester.c:
- Removed unreferenced variables: orderIndex, index, previousBalance, profit, lastTradeIndex
- Added explicit (int) cast for double-to-int conversion (line 1065)
- Added explicit (int) casts for pInSettings DISABLE_COMPOUNDING parameter (lines 2203, 2212)

AsirikuyCommon/Broker-tz.c:
- Added explicit (int) cast for size_t-to-int conversion in maxBytes assignment (line 82)

SymbolAnalyzer/SymbolAnalyzer.c:
- Added explicit (int) cast for pointer arithmetic result (line 447)
- Added explicit (int) casts for strlen() return values (lines 494-495)

Result: All core projects now build with significantly reduced warnings

+++ core/AsirikuyCommon/src/Broker-tz.c
@@ -79,7 +79,7 @@ AsirikuyReturnCode getTimezoneInfo(const char* pName, TimezoneInfo** pTimezoneIn
   if(nameLen > 0) {
     char hexBuffer[256] = "";
     char *p = hexBuffer;
-    int maxBytes = (nameLen < 10 ? nameLen : 10);
+    int maxBytes = (int)(nameLen < 10 ? nameLen : 10);
     for(i = 0; i < maxBytes && (p - hexBuffer) < (sizeof(hexBuffer) - 20); i++) {
       p += snprintf(p, sizeof(hexBuffer) - (p - hexBuffer), "0x%02x('%c') ", 
                      (unsigned char)pName[i], (pName[i] >= 32 && pName[i] < 127) ? pName[i] : '?');
